+++ R/utils.R
@@ -8,6 +8,7 @@ inter2inst <- function(x) {
 
 length_at_age <- function(population, age) {
   with(population, {
+    if(L2 < 0) L2 <- Linf * (1 - exp(-k * (-L2 - t0)))
     L <- Linf * (1 - exp(-k * (age - t0)))
     t2 <- -log(1 - min(L2/Linf, 1)) / k - t0
     L[age > t2] <- L2 + (Linf2 - L2) * (1 - exp(-k2 * (age[age > t2] - t2)))